+++ TicTacToe-Client/Scripts/TicTacToe.js
@@ -2,12 +2,24 @@
 // 3x3 Multidimensional Game array
 var gameArray = new Array(3);
 
+// Thickness of border for individual square on GameBoard
+var squareBorderThickness = 0;
+
+// Width of individual square on GameBoard
+var squareWidth = 0;
+
 // Which player starts the game?
 var player = "circle";
 
 $(document).ready(function () {
+    // Retrieve first game board square
+    var objGameBoardSquare = $('#square_0_0')[0]; // Native DOM element from jquery get()
+
+    squareBorderThickness = getSquareLineThickness(objGameBoardSquare);
+    squareWidth = objGameBoardSquare.scrollWidth - squareBorderThickness;
+
     generateGameArray(gameArray);
-    registerSquareClickEvents(gameArray);
+    registerSquareClickEvents(gameArray, squareWidth);
 
     $('#btnTest').on('click', function () {
         test(gameArray);
@@ -21,7 +33,7 @@ function generateGameArray(cubicArray) {
     }
 }
 
-function registerSquareClickEvents(cubicArray) {
+function registerSquareClickEvents(cubicArray, registerSquareWidth) {
     // GameBoard is a 3x3 matrix
     // Rows
     for (var i = 0; i < cubicArray.length; i++) {
@@ -31,7 +43,7 @@ function registerSquareClickEvents(cubicArray) {
 
             // Register onclick event
             $(refSquareName).on('click', function () {
-                squareClick(this);
+                clickSquare(this, registerSquareWidth);
             });
         }
     }
@@ -56,25 +68,7 @@ function test(cubicArray) {
     alert("Game array: \n" + arrayString);
 }
 
-function squareClick(object) {
-    //// Get style (for the clicked square)
-    //var squareStyle = object.attributes["style"].value;
-    //var styleArray = squareStyle.split(";");
-
-    //// Get line width of the path (for the clicked square)
-    //var lineWidth = "";
-    //for (var i = 0; i < styleArray.length; i++) {
-    //    var attributeArray = styleArray[i].split(':');
-    //    if (attributeArray[0].trim() == "stroke-width") {
-    //        lineWidth = Number(attributeArray[1].trim());
-    //    }
-    //}
-
-    var lineWidth = getSquareLineWidth(object);
-
-    // Get square width (for the clicked square)
-    var squareWidth = object.scrollWidth - lineWidth;
-
+function clickSquare(object, clickSquareWidth) {
     // Get [row][column] info (for the clicked square)
     var objectNameArray = object.id.split("_");
     var row = objectNameArray[1];
@@ -88,13 +82,13 @@ function squareClick(object) {
         // Current player?
         if (player == "circle") {
             // Add object to GameBoard
-            addCircle(row, column, squareWidth);
+            addCircle(row, column, clickSquareWidth);
 
             // Switch player
             player = "cross";
         } else if (player == "cross") {
             // Add object to GameBoard
-            addCross(row, column, squareWidth);
+            addCross(row, column, clickSquareWidth);
 
             // Switch player
             player = "circle";
@@ -103,28 +97,28 @@ function squareClick(object) {
         }
 
         // Victory?
-        checkVictory(gameArray);
+        checkVictory(gameArray, clickSquareWidth);
     }
 }
 
-function getSquareLineWidth(gameBoardSquare) {
+function getSquareLineThickness(gameBoardSquare) {
     // Get style (for the clicked square)
     var squareStyle = gameBoardSquare.attributes["style"].value;
     var styleArray = squareStyle.split(";");
 
-    // Get line width of the path (for the referenced GameBoard square)
-    var squareLineWidth = 0;
+    // Get line thickness of the path (for the referenced GameBoard square)
+    var borderThickness = 0;
     for (var i = 0; i < styleArray.length; i++) {
         var attributeArray = styleArray[i].split(':');
         if (attributeArray[0].trim() == "stroke-width") {
-            squareLineWidth = Number(attributeArray[1].trim());
+            borderThickness = Number(attributeArray[1].trim());
         }
     }
 
-    return squareLineWidth;
+    return borderThickness;
 }
 
-function checkVictory(cubicArray) {
+function checkVictory(cubicArray, victorySquareWidth) {
     var result = false;
     var lineData = "";
 
@@ -133,53 +127,61 @@ function checkVictory(cubicArray) {
     // Row 1
     lineData = cubicArray[0][0] + "," + cubicArray[0][1] + "," + cubicArray[0][2];
     if (lineData == "circle,circle,circle" || lineData == "cross,cross,cross") {
+        victoryLine(victorySquareWidth, "row1");
         result = true;
     }
 
     // Row 2
     lineData = cubicArray[1][0] + "," + cubicArray[1][1] + "," + cubicArray[1][2];
     if (lineData == "circle,circle,circle" || lineData == "cross,cross,cross") {
+        victoryLine(victorySquareWidth, "row2");
         result = true;
     }
 
     // Row 3
     lineData = cubicArray[2][0] + "," + cubicArray[2][1] + "," + cubicArray[2][2];
     if (lineData == "circle,circle,circle" || lineData == "cross,cross,cross") {
+        victoryLine(victorySquareWidth, "row3");
         result = true;
     }
 
     // Column 1
     lineData = cubicArray[0][0] + "," + cubicArray[1][0] + "," + cubicArray[2][0];
     if (lineData == "circle,circle,circle" || lineData == "cross,cross,cross") {
+        victoryLine(victorySquareWidth, "column1");
         result = true;
     }
 
     // Column 2
     lineData = cubicArray[0][1] + "," + cubicArray[1][1] + "," + cubicArray[2][1];
     if (lineData == "circle,circle,circle" || lineData == "cross,cross,cross") {
+        victoryLine(victorySquareWidth, "column2");
         result = true;
     }
 
     // Column 3
     lineData = cubicArray[0][2] + "," + cubicArray[1][2] + "," + cubicArray[2][2];
     if (lineData == "circle,circle,circle" || lineData == "cross,cross,cross") {
+        victoryLine(victorySquareWidth, "column3");
         result = true;
     }
 
     // "Backslash" Diagonal ("\")
     lineData = cubicArray[0][0] + "," + cubicArray[1][1] + "," + cubicArray[2][2];
     if (lineData == "circle,circle,circle" || lineData == "cross,cross,cross") {
+        victoryLine(victorySquareWidth, "backSlash");
         result = true;
     }
 
     // "Forward Slash" Diagonal ("/")
     lineData = cubicArray[0][2] + "," + cubicArray[1][1] + "," + cubicArray[2][0];
     if (lineData == "circle,circle,circle" || lineData == "cross,cross,cross") {
+        victoryLine(victorySquareWidth, "forwardSlash");
         result = true;
     }
 
     if (result == true) {
-        alert("Victory!");
+        //alert("Victory!");
         //test(gameArray);
     }
 }
@@ -188,15 +190,20 @@ function checkVictory(cubicArray) {
 function addCircle(row, col, squareSideLength) {
     // Draw SVG circle
     var objName = "circle" + "_" + row + "_" + col;
+    var objCircle = document.createElementNS("http://www.w3.org/2000/svg", "circle");
+
+    // Center
     var centerX = col * squareSideLength + 0.5 * squareSideLength;
     var centerY = row * squareSideLength + 0.5 * squareSideLength;
-    var objCircle = document.createElementNS("http://www.w3.org/2000/svg", "circle");
+
+    // Attributes
     objCircle.setAttribute("id", objName);
     objCircle.setAttribute("style", "stroke:blue;fill:none;");
     objCircle.setAttribute("cx", centerX);
     objCircle.setAttribute("cy", centerY);
     objCircle.setAttribute("r", 0.4 * squareSideLength);
 
+    // Draw object on SVG GameBoard
     $('#svgGameBoard').append(objCircle);
 }
 
@@ -223,31 +230,115 @@ function addCross(row, col, squareSideLength) {
     var pathForwardSlash = "m " + startForwardSlashX + " " + startForwardSlashY + " l -" + distanceX + " " + distanceY;
 
     // Cross
-    var pathCross = pathBackSlash + " " + pathOriginReturn + " " + pathForwardSlash;
+    var strPath = pathBackSlash + " " + pathOriginReturn + " " + pathForwardSlash;
 
+    // Attributes
     objCross.setAttribute("id", objName);
     objCross.setAttribute("style", "stroke:red;fill:none;");
-    //objCross.setAttribute("d", pathBackSlash);
-    //objCross.setAttribute("d", pathForwardSlash);
-    objCross.setAttribute("d", pathCross);
+    objCross.setAttribute("d", strPath);
 
+    // Draw object on SVG GameBoard
     $('#svgGameBoard').append(objCross);
 }
 
-//function victoryCrossOut(row, col, squareSideLength, orientation) {
-//    switch (orientation) {
-//        case "row1":
-//            break;
-//        case "column1":
-//            break;
-//        case "backSlash":
-//            var startX = 0.15 * squareSideLength;
-//            var startY = 0.15 * squareSideLength;
-//            var crossoutBeamDistance = 2.8 * squareSideLength;
-//            objCross.setAttribute("d", "m " + startX + " " + startY + " l " + crossoutBeamDistance + " " + crossoutBeamDistance);
-//            break;
-//        case "forwardSlash":
-//            break;
-//        default:
-//    }
-//}
+function victoryLine(squareSideLength, orientation) {
+    var objName = "victoryLine";
+    var offsetEdge = 0.1;
+    var offsetAxis = 0.2;
+    var objVictory = document.createElementNS("http://www.w3.org/2000/svg", "path");
+
+    switch (orientation) {
+        case "row1":
+            // Starting Point
+            var startX = offsetEdge * squareSideLength;
+            var startY = (0.5 - offsetAxis) * squareSideLength;
+
+            // Distance (across 3 squares)
+            var distanceX = 3 * squareSideLength - 2 * offsetEdge * squareSideLength;
+            var distanceY = (0.0 + 2 * offsetAxis) * squareSideLength;
+            break;
+
+        case "row2":
+            // Starting Point
+            var startX = offsetEdge * squareSideLength;
+            var startY = 1 * squareSideLength + (0.5 - offsetAxis) * squareSideLength;
+
+            // Distance (across 3 squares)
+            var distanceX = 3 * squareSideLength - 2 * offsetEdge * squareSideLength;
+            var distanceY = (0.0 + 2 * offsetAxis) * squareSideLength;
+            break;
+
+        case "row3":
+            // Starting Point
+            var startX = offsetEdge * squareSideLength;
+            var startY = 2 * squareSideLength + (0.5 - offsetAxis) * squareSideLength;
+
+            // Distance (across 3 squares)
+            var distanceX = 3 * squareSideLength - 2 * offsetEdge * squareSideLength;
+            var distanceY = (0.0 + 2 * offsetAxis) * squareSideLength;
+            break;
+
+        case "column1":
+            // Starting Point
+            var startX = 0 * squareSideLength + (0.5 - offsetAxis) * squareSideLength;
+            var startY = offsetEdge * squareSideLength;
+
+            // Distance (across 3 squares)
+            var distanceX = 2 * offsetAxis * squareSideLength;
+            var distanceY = 3 * squareSideLength - 2 * offsetEdge * squareSideLength;
+            break;
+
+        case "column2":
+            // Starting Point
+            var startX = 1 * squareSideLength + (0.5 - offsetAxis) * squareSideLength;
+            var startY = offsetEdge * squareSideLength;
+
+            // Distance (across 3 squares)
+            var distanceX = 2 * offsetAxis * squareSideLength;
+            var distanceY = 3 * squareSideLength - 2 * offsetEdge * squareSideLength;
+            break;
+
+        case "column3":
+            // Starting Point
+            var startX = 2 * squareSideLength + (0.5 - offsetAxis) * squareSideLength;
+            var startY = offsetEdge * squareSideLength;
+
+            // Distance (across 3 squares)
+            var distanceX = 2 * offsetAxis * squareSideLength;
+            var distanceY = 3 * squareSideLength - 2 * offsetEdge * squareSideLength;
+            break;
+
+        case "backSlash":
+            // Starting Point
+            var startX = offsetEdge * squareSideLength;
+            var startY = 2 * offsetAxis * squareSideLength;
+
+            // Distance (across 3 squares)
+            var distanceX = 3 * squareSideLength - 2 * offsetEdge * squareSideLength;
+            var distanceY = 3 * squareSideLength - 4 * offsetAxis * squareSideLength;
+            break;
+
+        case "forwardSlash":
+            // Starting Point
+            var startX = 3 * squareSideLength - offsetEdge * squareSideLength;
+            var startY = 2 * offsetAxis * squareSideLength;
+
+            // Distance (across 3 squares)
+            var distanceX = -1 * (3 * squareSideLength - 2 * offsetEdge * squareSideLength);
+            var distanceY = 3 * squareSideLength - 4 * offsetAxis * squareSideLength;
+            break;
+
+        default:
+    }
+
+    // Path
+    var strPath = "m " + startX + " " + startY + " l " + distanceX + " " + distanceY;
+
+    // Attributes
+    objVictory.setAttribute("id", objName);
+    objVictory.setAttribute("style", "stroke-width:15;stroke:black;fill:none;opacity:1.0");
+    objVictory.setAttribute("d", strPath);
+
+    // Draw object on SVG GameBoard
+    $('#svgGameBoard').append(objVictory);
+}